+++ src/client/components/bracket/Bracket.tsx
@@ -1,28 +1,27 @@
-import {
-  useState,
-  useEffect,
-  useLayoutEffect,
-  useReducer,
-  Reducer,
-} from 'react';
+import { useState, useEffect, useLayoutEffect, useReducer } from 'react';
 import testTournamentData from '../../../assets/test_data/test-tournament';
 import RoundColumn from '../RoundColumn';
-import changeDisplay from './reducer';
+import toggleView from './reducer';
 import axios from 'axios';
-import { MatchUpType, displayStateProps } from '../../../types';
+import { MatchUpType } from '../../../types';
 import processMatchups from './processMatchups';
-import initialDisplayState from './initialDisplayState';
 
-const TEST_TOURNAMENT_ID = '64863e1bf5a5d7a132318e76';
-
-export interface displayReducerProps {
-  type: string;
-  payload: any;
-}
+const initialDisplayState = {
+  unidirectional: true,
+  numberOfColumns: Math.log2(testTournamentData.matchUps.length + 1),
+  displaySettings: {
+    gridTemplateColumns: `repeat(${Math.log2(
+      testTournamentData.matchUps.length + 1
+    )}, 1fr)`,
+    columnGap: '10%',
+  },
+};
 
 const Bracket = () => {
+  // combine state updates with useReducer?
+
   const [displayState, displayDispatch] = useReducer(
-    changeDisplay,
+    toggleView,
     initialDisplayState
   );
 
@@ -34,7 +33,7 @@ const Bracket = () => {
   const [displayVotes, setDisplayVotes] = useState<boolean>(false);
   const [round, setRound] = useState<number>(1);
 
-  // //useCallback w/active tournament ID?
+  // //useCallback?
 
   const getMatchUps = async (id: string) => {
     console.log('GET');
@@ -54,11 +53,13 @@ const Bracket = () => {
   // but also the server has to get involved at some point to advanced contestants
   // hard-code test tournament id for now
   useEffect(() => {
-    getMatchUps(TEST_TOURNAMENT_ID);
+    getMatchUps('64863e1bf5a5d7a132318e76');
   }, []);
 
   useEffect(() => {
     console.log('USEEFFECT');
+    if (matchUpResponse.length) setIsLoading(false);
+
     // better to use array or object?
     const selectionArray = [];
     matchUpResponse.filter((el) => {
@@ -66,17 +67,17 @@ const Bracket = () => {
     });
   }, [matchUpResponse]);
 
-  // convert response from DB into array of arrays of matchup objects -- one sub-array per displayed column
+  // create matchUps object whose keys are round numbers and whose values are the array of matchups for each column
   useLayoutEffect(() => {
-    console.log('ULE');
-    if (matchUpResponse.length) setIsLoading(false);
+    console.log('USELAYOUTEFFECT');
+    // sorting logic moves to here
+    // arrays? destructuring?
+    // add column keys in render function
     const matchUpData = processMatchups(matchUpResponse, displayState);
     setMatchUps(matchUpData);
   }, [displayState, matchUpResponse]);
 
-  return isLoading ? (
-    <div>Loading...</div>
-  ) : (
+  return (
     <div>
       <div className='bracket-render-grid' style={displayState.displaySettings}>
         {matchUps.map((column, index) => {
@@ -85,19 +86,7 @@ const Bracket = () => {
           );
         })}
       </div>
-      <button
-        onClick={() =>
-          displayDispatch({
-            type: 'updateDisplay',
-            payload: {
-              unidirectional: !displayState.unidirectional,
-              matchUps: matchUpResponse.length,
-            },
-          })
-        }
-      >
-        Toggle View
-      </button>
+      <button onClick={() => displayDispatch('toggleView')}>Toggle View</button>
       <button onClick={() => setRound(() => round + 1)}>
         TEST: Next Round
       </button>
