+++ src/client/components/bracket/reducer.ts
@@ -1,34 +1,29 @@
 import { displayStateProps } from '../../../types';
 import testTournamentData from '../../../assets/test_data/test-tournament';
-import { displayReducerProps } from './Bracket';
 
-const changeDisplay = (
-  state: displayStateProps,
-  action: displayReducerProps
+const toggleView = (
+  props: displayStateProps,
+  action: string
 ): displayStateProps => {
-  const { type, payload } = action;
-
-  console.log('DISPATCH action ' + type);
-
-  if (type === 'updateDisplay') {
-    const newProps: displayStateProps = Object.assign(state, payload);
-    const newDisplaySettings = Object.assign({}, newProps.displaySettings);
-    console.log(newProps.numberOfColumns);
-    if (payload.unidirectional) {
-      newProps.numberOfColumns = Math.log2(payload.matchUps + 1);
+  console.log('DISPATCH action ' + action);
+  const newProps = Object.assign({}, props);
+  if (action === 'toggleView') {
+    if (props.unidirectional) {
+      newProps.unidirectional = false;
+      newProps.numberOfColumns =
+        Math.log2(testTournamentData.matchUps.length + 1) * 2 - 1;
     } else {
-      newProps.numberOfColumns = Math.log2(payload.matchUps + 1) * 2 - 1;
+      newProps.unidirectional = true;
+      newProps.numberOfColumns = Math.log2(
+        testTournamentData.matchUps.length + 1
+      );
     }
-    console.log(newProps.numberOfColumns);
-
-    newDisplaySettings.gridTemplateColumns = `repeat(${newProps.numberOfColumns}, 1fr)`;
-    newDisplaySettings.columnGap = payload.unidirectional ? '5%' : '10%';
-
-    newProps.displaySettings = newDisplaySettings;
-    console.log('NP: ', newProps);
-    return newProps;
   }
-  return state;
+  const newDisplaySettings = Object.assign({}, newProps.displaySettings);
+  newDisplaySettings.gridTemplateColumns = `repeat(${newProps.numberOfColumns}, 1fr)`;
+  newDisplaySettings.columnGap = props.unidirectional ? '5%' : '10%';
+  newProps.displaySettings = newDisplaySettings;
+  return newProps;
 };
 
-export default changeDisplay;
+export default toggleView;
